+++ src/components/_modals/DepositModal/index.tsx
@@ -6,33 +6,25 @@ import {
   TabPanel,
 } from "@chakra-ui/react"
 import { VFC } from "react"
-import { Token as TokenType } from "data/tokenConfig"
 
-interface FormValues {
-  depositAmount: number
-  slippage: number
-  selectedToken: TokenType
-}
 import { BaseModal } from "../BaseModal"
 import { SommelierTab } from "./SommelierTab"
 import React from "react"
 import { strategyPageContentData } from "data/strategyPageContentData"
 import { useRouter } from "next/router"
-import { ExchangeTab } from "./ExchangeTab"
-import { CustomTab } from "../CustomTab"
+import { CustomTab } from "../../_tabs/CustomTab"
+import { ExchangeTab } from "components/_tabs/ExchangeTab"
 
 type DepositModalProps = Pick<ModalProps, "isOpen" | "onClose">
 
 export const DepositModal: VFC<DepositModalProps> = (props) => {
   const id = useRouter().query.id as string
-  const uniswap = strategyPageContentData[id].exchange.find(
-    (v) => v.name === "Uniswap"
-  )
+  const exchanges = strategyPageContentData[id].exchange.length
 
   return (
     <BaseModal heading="Buy" {...props}>
       <Tabs variant="unstyled" isFitted>
-        {uniswap && (
+        {exchanges && (
           <TabList gap={2}>
             <CustomTab>Sommelier</CustomTab>
             <CustomTab>Exchange</CustomTab>
@@ -42,7 +34,7 @@ export const DepositModal: VFC<DepositModalProps> = (props) => {
           <TabPanel px={0}>
             <SommelierTab {...props} />
           </TabPanel>
-          {uniswap && (
+          {exchanges && (
             <TabPanel px={0} pt={8}>
               <ExchangeTab />
             </TabPanel>
