+++ src/components/_modals/WithdrawModal/index.tsx
@@ -17,20 +17,20 @@ import { BaseModal } from "../BaseModal"
 import React from "react"
 import { strategyPageContentData } from "data/strategyPageContentData"
 import { useRouter } from "next/router"
-import { ExchangeTab } from "./ExchangeTab"
 import { SommelierTab } from "./SommelierTab"
-import { CustomTab } from "../CustomTab"
+import { CustomTab } from "../../_tabs/CustomTab"
+import { ExchangeTab } from "components/_tabs/ExchangeTab"
 
 type WithdrawModalProps = Pick<ModalProps, "isOpen" | "onClose">
 
 export const WithdrawModal: VFC<WithdrawModalProps> = (props) => {
   const id = useRouter().query.id as string
-  const buyUrl = strategyPageContentData[id]?.buyUrl
+  const exchanges = strategyPageContentData[id]?.exchange.length
 
   return (
     <BaseModal heading="Sell" {...props}>
       <Tabs variant="unstyled" isFitted>
-        {buyUrl && (
+        {exchanges && (
           <TabList gap={2}>
             <CustomTab>Sommelier</CustomTab>
             <CustomTab>Exchange</CustomTab>
@@ -40,7 +40,7 @@ export const WithdrawModal: VFC<WithdrawModalProps> = (props) => {
           <TabPanel px={0}>
             <SommelierTab {...props} />
           </TabPanel>
-          {buyUrl && (
+          {exchanges && (
             <TabPanel px={0} pt={8}>
               <ExchangeTab />
             </TabPanel>
